requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.3
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
